requirements.txt
numpy>=1.26
scipy>=1.13
pandas>=2.2
pyyaml>=6.0
PyQt6>=6.7
pytest>=8.0